+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased changes
 - Removed unnecessary  `amount` parameter from `InvokeInstanceRequest`.
 - Added utility functions for converting between `CCDAmount` and `Energy`. Present in utility class `Converter`.
+- Added `Web3IdProof` class with `getWeb3IdProof` method to create Presentations. (And supporting classes)
 
 ## 7.0.0
 - Make the `energy` parameter for invoking an instance `Optional`.